+++ src/main/java/com/example/parking/domain/review/service/ReviewDomainService.java
@@ -1,6 +1,6 @@
 package com.example.parking.domain.review.service;
 
-import com.example.parking.domain.common.Association;
+import com.example.parking.support.Association;
 import com.example.parking.domain.member.Member;
 import com.example.parking.domain.parking.Parking;
 import com.example.parking.domain.review.Content;
